pr/LLM-workshop/requirements.txt
fastapi
uvicorn
mistralai
sentence-transformers
faiss-cpu
python-dotenv
pydantic
requests
PyMuPDF

